fix: potential update to old version of resource
Pulling the finalizer existence check out of the addFinalizer function
avoids the scenario where the resource is updated (to add the finalizer)
but later operations expect to be working with the latest version of the
resource. Making this check explicit inside of the reconcile loop allows
us to requeue the resource after adding the finalizer allowing later
operations to work with the latest version of the resource (without
having to re-get it).